requirements.txt
qwen-vl-utils>=0.0.8
torchvision
transformers>=4.46.3
accelerate>=1.1.1
huggingface_hub
spaces
gradio>=5.6.0